python-playground/pyproject.toml
[project]
name = "python-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.10.0",
]
